Quantized FP32 to Int4 with OS_IS_YX_OSV32_ISV2 Format
This commit introduces functions to convert quantized FP32 values to int4 using fp16 scales, with a decompression group size of 128.
Additionally, the function reorders data into OS_IS_YX_OSV32_ISV2 format after quantization.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/tensor/cl_operations/blas_kernels.cpp
@@ -76,7 +76,8 @@ void gemv_int4_cl(char *weight, uint16_t *scale, uint16_t *input,
   }
 
   /// @todo synchronize when only needed
-  blas_cc->command_queue_inst_.enqueueSVMMap(output, N * sizeof(float), true);
+  blas_cc->command_queue_inst_.enqueueSVMMap(output, N * sizeof(uint16_t),
+                                             true);
   if (!result) {
     throw std::runtime_error(
       "Failed to read output data for fully_connected_gpu_int4_gemv");
